Reverting back to original age-specific death rate
Reverting code back to using the age-specific death rate, where the death rate for a given age group is specific to the year listed in the same row in the abridged life tables (i.e. for the years 2020-2025 in the life tables, the death rates for 1-4 year-olds during those same years are listed in those rows).