Update to 2019 datasets
Minor change to vaccinate() since it wasn't running, updated to use the 2019 touchstone data.  Made the calibration plots a standard piece of the output.  Added a few helper functions "De"-suffixed, that run on the De suffixed R files.  These files keep an 11th column: the number of deaths in each time period.  They're not used in general modeling, but can be helpful in debugging/investigating the model.